# src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "0.1.0",
    "title": "DataPorts API with swagger",
    "description": "This is a simple REST API made with Node.JS and documented with Swagger",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "contact": {
      "name": "José A. Clemente",
      "email": "jclemente@prodevelop.es"
    }
  },
  "basePath": "/dac",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/images/{id}/template": {
      "get": {
        "tags": [
          "Images"
        ],
        "operationId": "getTemplate",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "bearerToken": []
          }
        ],
        "summary": "Returns a template for creating an agent",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the image that we want to retrieve its template",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "succesful operation",
            "schema": {
              "$ref": "#/definitions/GenericAnswer"
            }
          },
          "404": {
            "description": "Generic error",
            "schema": {
              "$ref": "#/definitions/GenericMessage"
            }
          }
        }
      }
    },
...
    "/orion/{id}": {
      "delete": {
        "tags": [
          "ORION"
        ],
        "summary": "Deletes a NGSI entity",
        "security": [
          {
            "bearerToken": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the entity that we want to delete",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "succesful operation",
            "schema": {
              "$ref": "#/definitions/GenericAnswer"
            }
          },
          "404": {
            "description": "Generic error",
            "schema": {
              "$ref": "#/definitions/GenericMessage"
            }
          }
        }
      }
    },
    "/orion/subscription/{id}": {
      "delete": {
        "tags": [
          "ORION"
        ],
        "summary": "Deletes a subscription ",
        "security": [
          {
            "bearerToken": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the subscription that we want to delete",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "succesful operation",
            "schema": {
              "$ref": "#/definitions/GenericAnswer"
            }
          },
          "404": {
            "description": "Generic error",
            "schema": {
              "$ref": "#/definitions/GenericMessage"
            }
          }
        }
      }
    },
    "/orion/getSubscriptions": {
      "get": {
        "tags": [
          "ORION"
        ],
        "summary": "Returns all the subscriptions available in the system",
        "security": [
          {
            "bearerToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "succesful operation",
            "schema": {
              "$ref": "#/definitions/GenericAnswer"
            }
          },
          "404": {
            "description": "Generic error",
            "schema": {
              "$ref": "#/definitions/GenericMessage"
            }
          }
        }
      }
    },
    "/orion/getAgentsAsociated": {
      "get": {
        "tags": [
          "ORION"
        ],
        "summary": "Returns agents associated to the entry parameters",
        "security": [
          {
            "bearerToken": []
          }
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Create an on demand agent",
            "required": true,
            "schema": {
              "$ref": "#/definitions/getAgents"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "succesful operation",
            "schema": {
              "$ref": "#/definitions/GenericAnswer"
            }
          },
          "404": {
            "description": "Generic error",
            "schema": {
              "$ref": "#/definitions/GenericMessage"
            }
          }
        }
      }
    },
    "/orion/{type}/entitiesByType": {
      "get": {
        "tags": [
          "ORION"
        ],
        "summary": "Recover entities by type",
        "security": [
          {
            "bearerToken": []
          }
        ],
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "required": true,
            "description": "type of the entity that we want to recover",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "succesful operation",
            "schema": {
              "$ref": "#/definitions/GenericAnswer"
            }
          },
          "404": {
            "description": "Generic error",
            "schema": {
              "$ref": "#/definitions/GenericMessage"
            }
          }
        }
      }
    },
    "/orion/createSubscriptions": {
      "post": {
        "tags": [
          "ORION"
        ],
        "summary": "Creates a subscription",
        "description": "Method called for creating subscriptions",
        "security": [
          {
            "bearerToken": []
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Create an on demand agent",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Subscription"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "succesful operation",
            "schema": {
              "$ref": "#/definitions/GenericAnswer"
            }
          },
          "404": {
            "description": "Generic error",
            "schema": {
              "$ref": "#/definitions/GenericMessage"
            }
          }
        }
      }
    },
    "/orion/createEntity": {
      "post": {
        "tags": [
          "ORION"
        ],
        "summary": "Creates an entity",
        "description": "Method called for creating entities",
        "security": [
          {
            "bearerToken": []
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Creates an entity in ORION",
            "required": true,
            "schema": {
              "$ref": "#/definitions/DataSource"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "succesful operation",
            "schema": {
              "$ref": "#/definitions/GenericAnswer"
            }
          },
          "404": {
            "description": "Generic error",
            "schema": {
              "$ref": "#/definitions/GenericMessage"
            }
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "bearerToken": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT token with the `Bearer` prefix, E.g., \"Bearer easkkassjejm.dadsaeea.ssdaee\""
    }
  },
  "definitions": {
...
    "DataSource": {
      "properties": {
        "Id": {
          "type": "string"
        },
        "image": {
          "type": "string"
        },
        "dataSource": {
          "type": "string"
        },
        "agentType": {
          "type": "string"
        }
      }
    },
    "GenericMessage": {
      "properties": {
        "status": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    },
    "GenericAnswer": {
      "properties": {
        "status": {
          "type": "string"
        },
        "message": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
...
    "getAgents": {
      "properties": {
        "agentType": {
          "type": "string"
        },
        "dataSourceId": {
          "type": "string"
        }
      }
    },
...
    "Subscription": {
      "properties": {
        "id": {
          "type": "string"
        },
        "type": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "notificationURL": {
          "type": "string"
        }
      }
    },
...
  }
}